refactor: removed redundant files
A lot of the files such as the app and bot files are not redundant with the new architecture I have delted them and updated the imports.
